refactor: Create dto for userController
