+++ api/src/main/java/grooteogi/dto/UserDto.java
@@ -1,0 +1,23 @@
+package grooteogi.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+public class UserDto {
+
+  @Data
+  @AllArgsConstructor
+  @NoArgsConstructor
+  public static class Password {
+    String password;
+
+  }
+
+  @Data
+  @Builder
+  public static class Response {
+
+  }
+}
