--- api/src/main/java/grooteogi/dto/user/ProfileDto.java
@@ -1,13 +1,0 @@
-package grooteogi.dto.user;
-
-import lombok.Getter;
-
-@Getter
-public class ProfileDto {
-
-  private String nickname;
-  private String imageUrl;
-  private String name;
-  private String address;
-  private String contact;
-}
