--- api/src/main/java/grooteogi/dto/user/PwDto.java
@@ -1,8 +1,0 @@
-package grooteogi.dto.user;
-
-import lombok.Getter;
-
-@Getter
-public class PwDto {
-  String password;
-}
